+++ attributes/default.rb
@@ -3,15 +3,19 @@
 # Licensed under Apache 2.0 - see the LICENSE file
 
 # Download/install
-default[:druid][:version] = "0.7.0"
-default[:druid][:mirror] = "http://static.druid.io/artifacts/releases"
-default[:druid][:druid_dir] = "druid-#{default[:druid][:version]}"
-default[:druid][:archive] = "#{default[:druid][:druid_dir]}-bin.tar.gz"
+default[:druid][:version] = '0.7.0'
+default[:druid][:mirror] = 'http://static.druid.io/artifacts/releases'
+default[:druid][:druid_dir] = 'druid-#{default[:druid][:version]}'
+default[:druid][:archive] = '#{default[:druid][:druid_dir]}-bin.tar.gz'
 default[:druid][:checksum] = "bd3f25b2c8711515c0c5786639bddee5019bd1655cc96b95559f183fbf08e853"
+
+default[:druid][:repository] = 'https://github.com/druid-io/druid.git'
+default[:druid][:revision] = 'e81ac2ba4302d488f6c9a3dda8a89af9c10d35e8' # Release 0.7.0
 
 # Installation
 default[:druid][:user] = "druid"
 default[:druid][:group] = "druid"
+default[:druid][:src_dir] = "/opt/druid_src"
 default[:druid][:install_dir] = "/opt/druid"
 default[:druid][:config_dir] = "/etc/druid"
 
@@ -34,5 +38,4 @@ default[:druid][:middleManager][:properties]["druid.port"] = 8086
 
 # Other
 default['java']['jdk_version'] = '7'
-