+++ CHANGELOG.md
@@ -6,6 +6,7 @@ druid CHANGELOG
 * Updated default Druid version to 0.7.0
 * Moved archive filename spec to attributes
 * Now use new common config file where appropriate
+* Now checkout druid source code from github and build using maven
 
 0.0.2
 -----